Refactored sender and put the UDP sender on a background thread.

+++ sender/flight_state.h
@@ -4,7 +4,7 @@
 extern int throttle_pct;
 extern int bank_angle;
 extern int pitch_angle;
-extern float altitude;
+extern int altitude;
 extern int rpm;
 
 #endif // FLIGHT_STATE_H
